requirements.txt
Flask==3.1.3
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
openpyxl==3.1.5
pandas==2.0.3
Werkzeug==3.1.3
python-dotenv==1.0.0
mysql-connector-python==8.4.0
requests==2.31.0
Pillow==10.0.0
email-validator==2.3.0
